# packages/api-contract/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/tmdb/movie/{movieId}": {
      "get": {
        "operationId": "TmdbController_getMovie_v1",
        "parameters": [
          {
            "name": "movieId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tmdb"
        ]
      }
    },
    "/v1/tmdb/debug/movie/{movieId}/raw": {
      "get": {
        "operationId": "TmdbController_getRawMovieResponse_v1",
        "parameters": [
          {
            "name": "movieId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tmdb"
        ]
      }
    },
    "/v1/tmdb/genres": {
      "get": {
        "operationId": "TmdbController_getMovieGenres_v1",
        "parameters": [
          {
            "name": "language",
            "required": false,
            "in": "query",
            "schema": {
              "example": "ko",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieGenresResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tmdb"
        ]
      }
    },
    "/v1/tmdb/discover": {
      "get": {
        "operationId": "TmdbController_discover_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieDiscoverResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tmdb"
        ]
      }
    },
    "/v1/tmdb/search": {
      "get": {
        "operationId": "TmdbController_search_v1",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "example": "인셉션",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieSearchResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tmdb"
        ]
      }
    },
...
    "/v1/kobis/movie/{movieCd}": {
      "get": {
        "operationId": "KobisController_getMovieInfo_v1",
        "parameters": [
          {
            "name": "movieCd",
            "required": true,
            "in": "path",
            "description": "KOBIS 영화 코드",
            "schema": {
              "example": "202",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "KOBIS 영화 상세 원본 JSON 조회",
        "tags": [
          "KOBIS"
        ]
      }
    },
    "/v1/kobis/movies/search": {
      "get": {
        "operationId": "KobisController_searchMovies_v1",
        "parameters": [
          {
            "name": "movieNm",
            "required": true,
            "in": "query",
            "description": "검색할 영화 이름",
            "schema": {
              "example": "디지몬 어드벤처",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "KOBIS 영화 이름 검색 원본 JSON 조회",
        "tags": [
          "KOBIS"
        ]
      }
    },
...
  },
  "info": {
    "title": "CINEMO API",
    "description": "Cinema in Motion — 영화관 로비 API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "MovieDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 550
          },
          "title": {
            "type": "string",
            "example": "파이트 클럽"
          },
          "original_title": {
            "type": "string",
            "example": "Fight Club"
          },
          "original_language": {
            "type": "string",
            "example": "en"
          },
          "overview": {
            "type": "string",
            "example": "한 남자가 반복되는 일상에서 벗어나기 위해..."
          },
          "poster_path": {
            "type": "string",
            "nullable": true,
            "example": "/poster.jpg"
          },
          "release_date": {
            "type": "string",
            "example": "1999-10-15"
          },
          "director": {
            "type": "string",
            "nullable": true,
            "example": "데이비드 핀처"
          },
          "cast": {
            "example": [
              "브래드 피트"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "trailerUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://www.youtube.com/watch?v=example"
          },
          "videoType": {
            "type": "string",
            "enum": [
              "trailer"
            ],
            "nullable": true
          },
          "genre_ids": {
            "example": [
              18,
              53
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "origin_countries": {
            "example": [
              "US"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "firstReleaseDate": {
            "type": "string",
            "nullable": true,
            "example": "1999-10-15"
          },
          "reReleaseDates": {
            "example": [
              "2026-09-16"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "title",
          "overview",
          "poster_path",
          "release_date",
          "director",
          "genre_ids",
          "origin_countries"
        ]
      },
      "MovieGenreDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 18
          },
          "name": {
            "type": "string",
            "example": "드라마"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "MovieGenresResponseDto": {
        "type": "object",
        "properties": {
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MovieGenreDto"
            }
          }
        },
        "required": [
          "genres"
        ]
      },
      "MovieDiscoverMovieDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 550
          },
          "adult": {
            "type": "boolean",
            "example": false
          },
          "title": {
            "type": "string",
            "example": "파이트 클럽"
          },
          "original_title": {
            "type": "string",
            "example": "Fight Club"
          },
          "original_language": {
            "type": "string",
            "example": "en"
          },
          "overview": {
            "type": "string",
            "example": "한 남자가 반복되는 일상에서 벗어나기 위해..."
          },
          "poster_path": {
            "type": "string",
            "nullable": true,
            "example": "/poster.jpg"
          },
          "release_date": {
            "type": "string",
            "example": "1999-10-15"
          },
          "genre_ids": {
            "example": [
              18,
              53
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "id",
          "adult",
          "title",
          "original_title",
          "original_language",
          "overview",
          "poster_path",
          "release_date",
          "genre_ids"
        ]
      },
      "MovieDiscoverResponseDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "total_pages": {
            "type": "number",
            "example": 1
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MovieDiscoverMovieDto"
            }
          }
        },
        "required": [
          "page",
          "total_pages",
          "results"
        ]
      },
      "MovieSearchItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 550
          },
          "title": {
            "type": "string",
            "example": "파이트 클럽"
          },
          "original_title": {
            "type": "string",
            "example": "Fight Club"
          },
          "original_language": {
            "type": "string",
            "example": "en"
          },
          "overview": {
            "type": "string",
            "example": "한 남자가 반복되는 일상에서 벗어나기 위해..."
          },
          "poster_path": {
            "type": "string",
            "nullable": true,
            "example": "/poster.jpg"
          },
          "release_date": {
            "type": "string",
            "example": "1999-10-15"
          }
        },
        "required": [
          "id",
          "title",
          "original_title",
          "original_language",
          "overview",
          "poster_path",
          "release_date"
        ]
      },
      "MovieSearchResponseDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "total_pages": {
            "type": "number",
            "example": 1
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MovieSearchItemDto"
            }
          }
        },
        "required": [
          "page",
          "total_pages",
          "results"
        ]
      },
...
    }
  }
}